@@ -14,12 +14,12 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
-        "silex/silex": "~1.0",
+        "php": ">=5.5.9",
+        "silex/silex": "^2.0",
         "twig/twig": "~1.12",
-        "symfony/twig-bridge": "~2.3",
-        "symfony/translation": "~2.3",
-        "knplabs/knp-paginator-bundle": "2.4.1",
+        "symfony/twig-bridge": "~2.3|~3.0",
+        "symfony/translation": "~2.3|~3.0",
+        "knplabs/knp-paginator-bundle": "~2.5",
         "cakephp/utility": "3.0.*@beta"
     },
     "require-dev": {
