@@ -1,8 +1,10 @@
 <?php
 namespace Tch\Silex;
 
-use Tch\Silex\Provider\PaginationServiceProvider;
 use Silex\Application;
+use Silex\Provider\TwigServiceProvider;
+use Silex\Provider\TranslationServiceProvider;
+use Tch\Silex\Provider\PaginationServiceProvider;
 use Symfony\Component\HttpFoundation\Request;
 
 class FunctionalTest extends \PHPUnit_Framework_TestCase
@@ -12,7 +14,9 @@ class FunctionalTest extends \PHPUnit_Framework_TestCase
 
     protected function setUp()
     {
-        $this->app = new Application();
+        $this->app = new Application(['locale' => 'en']);
+        $this->app->register(new TwigServiceProvider());
+        $this->app->register(new TranslationServiceProvider());
         $this->app->register(new PaginationServiceProvider());
         $this->app['knp_paginator.path'] = __DIR__ . '/../vendor/knplabs/knp-paginator-bundle';
 
