@@ -2,12 +2,16 @@
 namespace Tch\Silex\Provider;
 
 use Silex\Application;
+use Silex\Provider\TwigServiceProvider;
+use Silex\Provider\TranslationServiceProvider;
 
 class PaginationServiceProviderTest extends \PHPUnit_Framework_TestCase
 {
     public function test_register()
     {
-        $app = new Application();
+        $app = new Application(['locale' => 'en']);
+        $app->register(new TwigServiceProvider());
+        $app->register(new TranslationServiceProvider());
 
         $provider = new PaginationServiceProvider();
         $provider->register($app);
